requirements.txt
rich==14.2.0
python-nmap==0.7.1
requests==2.32.5